start work on bindings
