+++ src/compile/nodes/Element.ts
@@ -683,70 +683,6 @@ function getRenderStatement(
 	return `@createElement("${name}")`;
 }
 
-const events = [
-	{
-		eventNames: ['input'],
-		filter: (node: Element, name: string) =>
-			node.name === 'textarea' ||
-			node.name === 'input' && !/radio|checkbox|range/.test(node.getStaticAttributeValue('type'))
-	},
-	{
-		eventNames: ['change'],
-		filter: (node: Element, name: string) =>
-			node.name === 'select' ||
-			node.name === 'input' && /radio|checkbox/.test(node.getStaticAttributeValue('type'))
-	},
-	{
-		eventNames: ['change', 'input'],
-		filter: (node: Element, name: string) =>
-			node.name === 'input' && node.getStaticAttributeValue('type') === 'range'
-	},
-
-	{
-		eventNames: ['resize'],
-		filter: (node: Element, name: string) =>
-			dimensions.test(name)
-	},
-
-	// media events
-	{
-		eventNames: ['timeupdate'],
-		filter: (node: Element, name: string) =>
-			node.isMediaNode() &&
-			(name === 'currentTime' || name === 'played')
-	},
-	{
-		eventNames: ['durationchange'],
-		filter: (node: Element, name: string) =>
-			node.isMediaNode() &&
-			name === 'duration'
-	},
-	{
-		eventNames: ['play', 'pause'],
-		filter: (node: Element, name: string) =>
-			node.isMediaNode() &&
-			name === 'paused'
-	},
-	{
-		eventNames: ['progress'],
-		filter: (node: Element, name: string) =>
-			node.isMediaNode() &&
-			name === 'buffered'
-	},
-	{
-		eventNames: ['loadedmetadata'],
-		filter: (node: Element, name: string) =>
-			node.isMediaNode() &&
-			(name === 'buffered' || name === 'seekable')
-	},
-	{
-		eventNames: ['volumechange'],
-		filter: (node: Element, name: string) =>
-			node.isMediaNode() &&
-			name === 'volume'
-	}
-];
-
 function shouldHaveAttribute(
 	node,
 	attributes: string[],
