+++ src/compile/nodes/shared/Node.ts
@@ -138,26 +138,6 @@ export default class Node {
 		if (this.parent) return this.parent.findNearest(selector);
 	}
 
-	getOrCreateAnchor(block: Block, parentNode: string, parentNodes: string) {
-		// TODO use this in EachBlock and IfBlock — tricky because
-		// children need to be created first
-		const needsAnchor = this.next ? !this.next.isDomNode() : !parentNode || !this.parent.isDomNode();
-		const anchor = needsAnchor
-			? block.getUniqueName(`${this.var}_anchor`)
-			: (this.next && this.next.var) || 'null';
-
-		if (needsAnchor) {
-			block.addElement(
-				anchor,
-				`@createComment()`,
-				parentNodes && `@createComment()`,
-				parentNode
-			);
-		}
-
-		return anchor;
-	}
-
 	remount(name: string) {
 		return `${this.var}.m(${name}._slotted.default, null);`;
 	}
